requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
cohere==5.5.0